Use the model's integer widths in the networkd link specification
Type VLAN.ID, Section.Index, and Entry.Index as uint16, Address.PrefixLength
as uint8, and Family.RouteMetric and Delegation.RouterLifetimeSeconds as
uint32 in internal/networkd/spec.go, decode the matching wire leaves at those
widths in internal/networkjson/networkjson.go, and update the three test
files that build a specification.

+++ internal/networkd/spec_test.go
@@ -22,7 +22,7 @@ func staticLinkSpec(files []networkd.File) networkd.Spec {
 				Addresses:   []networkd.Address{{IP: netip.MustParseAddr("203.0.113.2"), PrefixLength: 29}},
 				DHCP:        new(false),
 				Gateway:     netip.MustParseAddr("203.0.113.1"),
-				RouteMetric: new(10),
+				RouteMetric: new(uint32(10)),
 			},
 		},
 		IPv6: &networkd.FamilyV6{
